+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # NEXT RELEASE
-* Expose TheAlgorithm.isLoading() method
+
+### v0.25.1
+* Expose `TheAlgorithm.isLoading()` method
+* `triggerFeedUpdate()` is no longer called recursively
 
 # v0.25.0
-* **[BREAKING CHANGE]** Rename TheAlgorithm.getFeed() to triggerFeedUpdate(). No longer returns anything.
-* **[BREAKING CHANGE]** Rename setFeedInApp param to setTimelineInApp
+* **[BREAKING CHANGE]** Rename `TheAlgorithm.getFeed()` to `triggerFeedUpdate()`. No longer returns anything.
+* **[BREAKING CHANGE]** Rename `setFeedInApp` param to `setTimelineInApp`
 * Simplify the alternate score dict
 * Disallow setting the catchup checkpoint before the overall app timeline cutoff
 * UserData object now just stores a lookup dict for `followedAccounts` instead of all the actual followed `Account` objects